Remove freetype2-demos-infinality-remix.
The freetype2-demos-infinality-remix package was built using a split
package with freetype2-infinality-remix. However, since
freetype2-infinality-remix is not installed until after the demos
build, the systems pre-existing freetype2 headers would be used for
the demos build. If the installed freetype version is >2.10.1, then
the 2.10.1 demos won't build due to header differences.

The solution is to not use a split package and split the freetype2
build from the demos build.

However, since there is no actual difference between the official
freetype2-demos package and ours, drop the package all together.

Fixes #18.