Disable fontconfig docs to fix broken build.
The fontconfig build broke due to changes in the docs generator. The official
Arch package is also broken in the same way. Since we track the Arch official
package, disable docs until the official package is updated and an upstream
fix is included.

Docs can still be installed via the fontconfig-docs package.

Fixes #19.